Extract expand_trait fn
